make WIP component shared and move content to constants
