+++ src/applications/simple-forms/26-4555/containers/App.jsx
@@ -5,13 +5,15 @@ import PropTypes from 'prop-types';
 import RoutedSavableApp from 'platform/forms/save-in-progress/RoutedSavableApp';
 import { toggleValues } from 'platform/site-wide/feature-toggles/selectors';
 import FEATURE_FLAG_NAMES from 'platform/utilities/feature-toggles/featureFlagNames';
-import formConfig from '../config/form';
 
-import { WIP } from '../components/WIP';
+import formConfig from '../config/form';
+import { workInProgressContent } from '../definitions/constants';
+
+import { WIP } from '../../shared/components/WIP';
 
 export function App({ location, children, show264555 }) {
   if (!show264555) {
-    return <WIP />;
+    return <WIP content={workInProgressContent} />;
   }
   return (
     <RoutedSavableApp formConfig={formConfig} currentLocation={location}>
